requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
BeautifulSoup==3.2.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.7
constantly==15.1.0
cryptography==2.1.4
deluge==1.3.15
enum34==1.1.6
Flask==0.8
Flask-Uploads==0.2.1
gunicorn==19.7.1
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
Markdown==2.6.10
MarkupSafe==1.1.0
micawber==0.3.4
numpy==1.13.3
PAM==0.4.2
peewee==2.10.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyBarcode==0.7
pycairo==1.16.2
pycrypto==2.6.1
pygame===1.9.1release
Pygments==2.2.0
pygobject==3.26.1
pyOpenSSL==17.5.0
pyserial==3.4
python-libtorrent==1.1.5
pyxdg==0.25
rebound-cli==1.1.0a1
requests==2.18.4
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
SQLAlchemy==1.2.0
svgwrite==1.1.2
Twisted==17.9.0
unity-lens-photos==1.0
urllib3==1.22
urwid==2.0.1
Werkzeug==0.14.1
wxPython==3.0.2.0
wxPython-common==3.0.2.0
zope.interface==4.3.2
